refactor : 코스 상세조회 평점 관련 로직 수정

+++ src/main/java/umc/catchy/domain/course/domain/Course.java
@@ -32,7 +32,7 @@ public class Course extends BaseTimeEntity {
 
     private boolean hasReview;
 
-    private Float rating;
+    private Double rating;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
